cloud_function/requirements.txt
functions-framework==3.*
google-cloud-language==2.9.1
google-cloud-bigquery
flask